Fix 32bit overflow in monitor.c (issue #715) (#716)
If number of probes is greater than 1, then any scan above 50%
will result in the multiplication in the given line to overflow.
This triggers a false minimum comparison, yielding wildly incorrect
scan time estimations.

The fix is to cast it to a uint64_t, as is done for the other clauses
above.

+++ src/monitor.c
@@ -150,7 +150,7 @@ double compute_remaining_time(double age, uint64_t packets_sent, uint64_t iterat
 		}
 		if (zsend.max_index) {
 			double done = (double)packets_sent /
-				      (zsend.max_index * zconf.packet_streams /
+				      ((uint64_t)zsend.max_index * zconf.packet_streams /
 				       zconf.total_shards);
 			remaining[4] =
 			    (1. - done) * (age / done) + zconf.cooldown_secs;
